+++ .gitignore
@@ -10,7 +10,6 @@
 .DS_Store
 /build
 /captures
-.externalNativeBuild
 .cxx
 local.properties
