+++ app/build.gradle
@@ -20,7 +20,17 @@ android {
     }
 
     buildTypes {
+        debug {
+            versionNameSuffix "_Debug"
+            debuggable true
+//            buildConfigField 'String', 'BASE_URL', baseUrl
+            resValue "string", "app_name_build", "SSLPinningDemoDebug"
+            //resValue("string", "google_maps_key", localProperties['google_maps_key_debug'])
+        }
         release {
+            debuggable false
+            minifyEnabled true
+            resValue "string", "app_name_build", "SSLPinningDemo"
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
@@ -42,7 +52,7 @@ dependencies {
 //    def lifecycle_version = "2.0.0"
     def material_version = "1.8.0"
     def retrofit_version = "2.9.0"
-    def glide_version = '4.10.0'
+    def glide_version = '4.11.0'
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
@@ -63,7 +73,7 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.1'
 
     implementation "com.github.bumptech.glide:glide:$glide_version"
-    annotationProcessor "com.github.bumptech.glide:compiler:$glide_version"
+//    annotationProcessor "com.github.bumptech.glide:compiler:$glide_version"
     kapt "com.github.bumptech.glide:compiler:$glide_version"
 
     // ViewModel
