+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
+        android:label="@string/app_name_build"
         android:supportsRtl="true"
         android:theme="@style/Theme.CertificatePinningDemo"
         tools:targetApi="31">
