+++ app/src/main/java/com/rogergcc/certificatepinningdemo/core/ObjetsProviders.kt
@@ -1,0 +1,7 @@
+package com.rogergcc.certificatepinningdemo.core
+
+import com.google.gson.Gson
+
+object ObjetsProviders {
+    val gson: Gson = Gson()
+}
