--- app/src/main/java/com/rogergcc/certificatepinningdemo/core/viewextensions.kt
@@ -1,24 +1,0 @@
-package com.rogergcc.certificatepinningdemo.core
-
-import android.content.Context
-import android.view.View
-import android.view.inputmethod.InputMethodManager
-import com.google.gson.Gson
-
-fun View.hide() {
-    this.visibility = View.GONE
-}
-
-fun View.show() {
-    this.visibility = View.VISIBLE
-}
-
-fun View.hideKeyboard() {
-    val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-    imm.hideSoftInputFromWindow(windowToken, 0)
-}
-
-
-object GsonProvider {
-    val gson: Gson = Gson()
-}
