+++ app/src/main/java/com/rogergcc/certificatepinningdemo/ui/common/viewextensions.kt
@@ -1,0 +1,57 @@
+package com.rogergcc.certificatepinningdemo.ui.common
+
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+import android.widget.ImageView
+import android.widget.Toast
+import androidx.annotation.DrawableRes
+import androidx.fragment.app.Fragment
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.engine.DiskCacheStrategy
+
+fun View.hide() {
+    this.visibility = View.GONE
+}
+
+fun View.show() {
+    this.visibility = View.VISIBLE
+}
+
+fun View.hideKeyboard() {
+    val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    imm.hideSoftInputFromWindow(windowToken, 0)
+}
+
+
+fun ImageView.loadUrl(url: String) {
+    Glide.with(context)
+        .load(url)
+        .centerCrop()
+        .into(this)
+}
+
+fun ImageView.loadImageFromResource(@DrawableRes resourceId: Int) {
+    Glide.with(context)
+        .load(resourceId)
+        .centerCrop()
+        .diskCacheStrategy(DiskCacheStrategy.ALL)
+        .into(this)
+}
+
+inline fun <T : View> T.showIf(condition: (T) -> Boolean) {
+    if (condition(this)) {
+        show()
+    } else {
+        hide()
+    }
+}
+
+fun Context.showToast(message: String, duration: Int = Toast.LENGTH_SHORT) {
+    Toast.makeText(this, message, duration).show()
+}
+
+fun Fragment.showToast(message: String, duration: Int = Toast.LENGTH_SHORT) {
+    requireContext().showToast(message, duration)
+}
+
