+++ app/src/main/res/layout/fragment_search_user.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/normal_padding"
-    tools:context=".ui.FirstFragment">
+    tools:context=".ui.SearchUserFragment">
 
     <LinearLayout
